move web-assets java code to "opennms-web-api"
