+++ core/test-api/rest/pom.xml
@@ -80,9 +80,5 @@
       <artifactId>json</artifactId>
       <version>${jsonVersion}</version>
     </dependency>
-    <dependency>
-      <groupId>org.opennms.core</groupId>
-      <artifactId>org.opennms.core.web-assets</artifactId>
-    </dependency>
   </dependencies>
 </project>
