+++ opennms-web-api/src/test/resources/assets/assets.json
