+++ opennms-web-api/src/test/resources/assets/assets.min.json
